backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
scikit-learn==1.4.0
scikit-image==0.22.0
scipy==1.12.0
python-dotenv==1.0.0
gunicorn==21.2.0
psutil==5.9.8
Werkzeug==3.0.1
